requirements.txt
fastapi
uvicorn
pydantic
langchain
langgraph
langchain-ollama
psycopg2-binary
python-dateutil
python-dotenv
mcp
gradio